% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_bda}
\alias{add_bda}
\title{Add a binary dichotomic algorithm (BDA).}
\usage{
add_bda(i_1, i_2, Q_11, Q_12, Q_21, Q_22)
}
\arguments{
\item{i_1}{A unsigend Integer from 1 - [WORD LENGTH]}

\item{i_2}{A unsigend Integer from 1 - [WORD LENGTH]}

\item{Q_11}{A single character. Has to be one of \{'A', 'C', 'G', 'T', 'U'\}}

\item{Q_12}{A single character. Has to be one of \{'A', 'C', 'G', 'T', 'U'\}}

\item{Q_21}{A single character. Has to be one of \{'A', 'C', 'G', 'T', 'U'\}}

\item{Q_22}{A single character. Has to be one of \{'A', 'C', 'G', 'T', 'U'\}}
}
\value{
Boolean value. True if the rules parameters are correct
}
\description{
Each added BDA results one binary diget. You can add multibe BDAs for one procedure. To add a rule, you need to start a BDA procedure first.
To do so you can eihter call \code{\link{code_start_bda}} or \code{\link{seq_start_bda}}. To start a procedure
for all codons call \code{\link{start_bda}}.
To add a BDA simply Call \code{\link{add_bda}}. To execute the BDAs call \code{\link{run_bda}}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
code_start_bda(c("ACG", "CAG"))
add_bda(1, 2, "C", "G", "A", "T")
res.binary <- run_bda()

}
