% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdaTools.R
\name{BDATools}
\alias{BDATools}
\title{BDA (binary dichotomic algorithm) Object}
\description{
This Object contains a binary dichotomic algorithm (BDA) procedure. Use \link{start_bda} or \link{code_start_bda} to get a BDATools object. \cr 
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}

seealso [start_bda] for all rna codons, [code_start_bda] for code, [add_rule] add a BDA,
          [run_bda] for executing and run the BDAs, [run_bda_as_matrix] for a matrix result
}
\examples{
bda_obj <- start_bda()
add_rule(bda_obj, 1, 2, "C", "G", "A", "U")
add_rule(bda_obj, 1, 3, "A", "G", "C", "U")
res <- run_bda_as_matrix(bda_obj)

}
