% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgorithm.R
\name{cluster_add_weights}
\alias{cluster_add_weights}
\title{Add weights to a Cluster Algorithm object}
\usage{
cluster_add_weights(ca_obj, pos, weight, origen.base = "", target.base = "")
}
\arguments{
\item{pos}{is an integer between 1 and code tuple size}

\item{origen.base}{is a string. The symbol at the position of the outgoing word (optional if target.base is not set)}

\item{target.base}{is a string. The symbol at the position of the ingoing word (optional)}

\item{code}{is a CodonClusteringAlgorithm object}
}
\description{
ToDo
}
\examples{
ca_obj <-  start_cluster_algorithm(c("AC", "01", "AA", "01", "CA", "11", "CC", "10"))
# XC -1+4-> XA
cluster_add_weights(ca_obj, 2, 4, "C", "A")

ca_obj <-  start_cluster_algorithm(c("AC", "01", "AA", "01", "CA", "11", "CC", "10"))
# XC -1+4-> XY
cluster_add_weights(ca_obj, 2, 4, "C")

ca_obj <-  start_cluster_algorithm(c("AC", "01", "AA", "01", "CA", "11", "CC", "10"))
# XY -1+4-> XZ
cluster_add_weights(ca_obj, 2, 4)

}
