% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgorithm.R
\name{cluster_set_matrix_weight}
\alias{cluster_set_matrix_weight}
\title{Add weights´matrix to a Cluster Algorithm object}
\usage{
cluster_set_matrix_weight(ca_obj, weight.matrix.list)
}
\arguments{
\item{weight.matrix.list}{the length of the list must be equal to the tuple size of the code. Each element must be a matrix of length}

\item{code}{is a CodonClusteringAlgorithm object}
}
\description{
ToDo
}
\examples{
ca_obj <-  start_cluster_algorithm(c("AC", "01", "AA", "01", "CA", "11", "CC", "10"))
# GC -2+4-> GA
w.m <- generate_weight_matrix()
w.m[[2]]["C", "A"] <- 4
w.m[[1]]["G", "G"] <- 2
cluster_set_matrix_weight(ca_obj, w.m)

}
