% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_as_vector}
\alias{code_as_vector}
\title{Returns a code as vector.}
\usage{
code_as_vector(code, tuple_length = -55555L)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
StringVector code as vector.
}
\description{
Turns a sequence or a single string code into a string vector.
}
\examples{
code_vec <- code_as_vector("ACGCAG", 3)
code_vec <- code_as_vector("ACG CAG")

}
