% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_block_length}
\alias{code_block_length}
\title{Block length}
\usage{
code_block_length(code, tuple_length = -55555L)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
block length
}
\description{
This function gets the block length of a code.
}
\examples{
block.length <- code_block_length(c("ACG", "GAT"))
block.length <- code_block_length("ACGGAT", tuple_length=3)
block.length <- code_block_length("ACG GAT")

}
