% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_check_if_comma_free}
\alias{code_check_if_comma_free}
\title{Check if a code is comma free.}
\usage{
code_check_if_comma_free(code)
}
\arguments{
\item{code}{is a DNA or RNA code as string vector.}
}
\value{
Boolean value. True if the code comma free.
}
\description{
The function checks if the code is comma free. 
Comma free is a stronger restricted version of the circular code property.
For more info on this subject read:\cr
\link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5492142/},\cr
\link{http://dpt-info.u-strasbg.fr/~c.michel/Circular_Codes.pdf},\cr
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_check_if_comma_free(c("ACG", "CAG"))

}
