% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_check_if_self_complementary}
\alias{code_check_if_self_complementary}
\title{Check if a code is self complementary.}
\usage{
code_check_if_self_complementary(code)
}
\arguments{
\item{code}{A vector with codons.}
}
\value{
Boolean value. True if the code self complementary.
}
\description{
The function checks if the code is self complementary. A self complementary code contains for 
any codon (word) in the code the corresponding anti-codons.
For more info on this subject read:\cr
\link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5492142/},\cr
\link{http://dpt-info.u-strasbg.fr/~c.michel/Circular_Codes.pdf},\cr
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_check_if_self_complementary(c("ACG", "CAG"))

}
