% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_check_if_tessera}
\alias{code_check_if_tessera}
\title{Tessera check function}
\usage{
code_check_if_tessera(code)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}
}
\value{
if code only contains tessera
}
\description{
This function checks if all words in the code are correct Tessera words.
}
\examples{
is_tessera <- code_check_if_tessera(c("ACGT", "GATC"))
is_tessera <- code_check_if_tessera("ACGTGATC" )
is_tessera <- code_check_if_tessera("ACGT GATC")

}
