% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPloter.R
\name{code_factor_cycle}
\alias{code_factor_cycle}
\title{Prepares a graph G(X) with only those edges which are parte of a cycle.}
\usage{
code_factor_cycle(code, tuple_length)
}
\arguments{
\item{code}{is either a string vector or a string. it can be a DNA or RNA sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
A igraph (<http://igraph.org/r/>) object: Edges and vertices of only the cycles of an graph representing a circular code.
}
\description{
Prepares a R igraph object. Extracts all circular paths of the Graph G(X).
If the graph shows no cycle the graph will be empty. Otherwise it returns a igraph object showing all
circular paths.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
G <- code_factor_cycle(c("ACG", "CAG"))
G <- code_factor_cycle("ACGCAG", tuple_length=3)
G <- code_factor_cycle("ACG CAG")
plot(G)

}
\seealso{
[seq_factor_circles(seq)] for sequence, [code_get_all_circle_as_vector()] for the results as vector
}
