% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_get_all_logest_path_as_vector}
\alias{code_get_all_logest_path_as_vector}
\title{Check if a code is self complementary.}
\usage{
code_get_all_logest_path_as_vector(code)
}
\arguments{
\item{code}{A vertor with codons.}
}
\value{
Boolean value. True if the code self complementary.
}
\description{
The function checks if the code is self complementary. A self complementary code contains for 
any codon (word) the corresponding anti-codons. 
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_check_if_self_complementary(c("ACG", "CAG"))

}
