% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_get_amino_acids}
\alias{code_get_amino_acids}
\title{Get amino acids encoded by a code}
\usage{
code_get_amino_acids(code, idx_trans_table = 1L)
}
\arguments{
\item{code}{is either a string vector or a string. It should be a RNA/DNA - code or a sequence.}

\item{idx_trans_table}{the index of a Genetic Code table as int. (check \link{print_all_translation_table})}
}
\value{
String vector. list of amino acids
}
\description{
Returns the amino acids encoded by a codes. The code can contain only CYTOSINE (C), ADENINE (A), GUANINE (G)
and THYMINE (T) or URACIL (U) bases. If no other translation table is selecte the function will use the 
\emph{standard genetic code}. A different tranlastion table has to be added by index. Therefore, (see \link{print_all_translation_tables})\cr
For more info on this subject read:\cr
\link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5492142/},\cr
\link{http://dpt-info.u-strasbg.fr/~c.michel/Circular_Codes.pdf},\cr
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_get_amino_acids(c("ACG", "CAG"), idx_trans_table=2)
code_get_amino_acids("ACGCAG", idx_trans_table=2)
code_get_amino_acids("ACG CAG", idx_trans_table=2)

}
