% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_get_one_circle_as_vector}
\alias{code_get_one_circle_as_vector}
\title{Finds one circle constructable in a code.}
\usage{
code_get_one_circle_as_vector(code)
}
\arguments{
\item{code}{A DNA or RNA code as string vector represented by the graph.}
}
\description{
Prepares a R path string vector. Extracts all circle path of the Graph G(X).
If the graph shows no cycle the vector will be empty. Otherwise it returns a vector with one
circle paths.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
l_path <- code_get_one_circle_as_vector(c("ACG", "CGA"))

}
\seealso{
[code_factor_circles(code)] for code, [code_get_all_circle_as_vector()] for the results as list
}
