% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_get_one_cycles_as_vector}
\alias{code_get_one_cycles_as_vector}
\title{Finds one circular sequences constructable in a code.}
\usage{
code_get_one_cycles_as_vector(code, tuple_length = -1L)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
List of nodes representing a circular path in G(X).
}
\description{
Prepares a R path string vector. Extracts all cycles in the Graph G(X).
If the graph has no cycles the vector will be empty. Otherwise it returns a vector with one the nodes of
circular path.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
l_graph <- code_get_one_cycles_as_vector(c("ACG", "CGA"))
l_graph <- code_get_one_cycles_as_vector("ACGCGA", 3)
l_graph <- code_get_one_cycles_as_vector("ACG CGA")

}
\seealso{
[code_factor_cycles()] for code, [code_get_all_cycles_as_vector()] for the results as list
}
