% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_is_translatable}
\alias{code_is_translatable}
\title{Is a code translatable?}
\usage{
code_is_translatable(code, length = -1L)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
Boolean value. True if a fitting translation table exists.
}
\description{
This function checks if a code can be translated into amino acids. Therefore, it simply checks if a translation table which fits the tuple size of the code
is in the system.\cr.
\link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5492142/},\cr
\link{http://dpt-info.u-strasbg.fr/~c.michel/Circular_Codes.pdf},\cr
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_is_translatable(c("ACG", "CAG"))
code_is_translatable("ACGCAG", 3)
code_is_translatable("ACG CAG")

}
