% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_named_transform_tuples}
\alias{code_named_transform_tuples}
\title{Transformation of all tuples}
\usage{
code_named_transform_tuples(trans_name, code, tuple_length = -55555L)
}
\arguments{
\item{trans_name}{tname of a transformation. listed in description.}

\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
transformed code as String vector
}
\description{
This function transforms all tuples in code. The single letters get transformed by the rules which are set as parameter.
The rules are the predefined and listed below. This only works for genetic gen codes and sequences\cr
}
\details{
SW = (A, T,C,G) -> (T, A, G,C)\cr
YR = (A, T,C,G) -> (G,C, T,A)\cr
KM = (A, T,C,G) -> (C, G, A, T)\cr
AT = (A, T,C,G) -> (T, A,C,G)\cr
CG = (A, T,C,G) -> (A, T, G,C)\cr
ACTG = (A, T,C,G) -> (C, G, T,A)\cr
AGTC = (A, T,C,G) -> (G,C, A, T)\cr
}
\examples{
transformed_tuples <- code_named_transform_tuples("I", c("ACG", "GAT"))
transformed_tuples <- code_named_transform_tuples("ACTG", "CAGT", "ACGGAT", tuple_length=3)
transformed_tuples <- code_named_transform_tuples("ACTG", "CAGT", "ACG GAT")

}
