% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_path_end_vertices_miner}
\alias{code_path_end_vertices_miner}
\title{Pathend vertices miner}
\usage{
code_path_end_vertices_miner(code, tuple_length = -55555L)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
list of vertices
}
\description{
This function finds all vertices which have no outgoing edges in the associated graph.
}
\examples{
vertices <- code_path_end_vertices_miner(c("ACG", "GAT"))
vertices <- code_path_end_vertices_miner("ACGGAT", tuple_length=3)
vertices <- code_path_end_vertices_miner("ACG GAT")

}
