% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PdfManager.R
\name{code_pdf_report}
\alias{code_pdf_report}
\title{Generate a PDF/latex report}
\usage{
code_pdf_report(code, tuple_length = -55555)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\description{
The function makes a dir called: "GCATR_Report" in your current working dir. In the following steps it cops a .Rnw template file
into the new dir. Afterwords it executes knitr with parameters based on the your input. All the generated files are in the new dir.
}
\examples{
code_pdf_report(c("ACG", "CAG"))
code_pdf_report("ACGCAG", 3)
code_pdf_report("ACG CAG")


}
