% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_strip_complements}
\alias{code_strip_complements}
\title{Code strip complements}
\usage{
code_strip_complements(code, tuple_length = -55555L)
}
\arguments{
\item{code}{is either a string vector or a string. It has to be a RNA/DNA - code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
a string the code reduced to a strongly comma-free code
}
\description{
This function removes one codon of each codon anti-codon pair. (see  \link{code_check_if_self_complementary})
}
