% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_transform_tuples}
\alias{code_transform_tuples}
\title{Transformation of all tuples}
\usage{
code_transform_tuples(from, to, code, tuple_length = -55555L)
}
\arguments{
\item{from}{the origin letters which are maped to the letters of the \emph{to} parameter.}

\item{to}{the transformation target letters which are maped letters of to the \emph{from} parameter.}

\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
transformed code as String vector
}
\description{
This function transforms all tuples in code. The single letters get transformed by the rules which are set as parameter.
The rules are defined as two strings, the \emph{from} and the \emph{to} parameter. These two parameters have to be
strings of the same length. Each letter in the \emph{from} string gets transformed to the corresponding letter at the same
position of the \emph{to} parameter.
}
\examples{
transformed_tuples <- code_transform_tuples("ACTG", "CAGT", c("ACG", "GAT"))
transformed_tuples <- code_transform_tuples("ACTG", "CAGT", "ACGGAT", tuple_length=3)
transformed_tuples <- code_transform_tuples("ACTG", "CAGT", "ACG GAT")

}
