% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_tuple_length}
\alias{code_tuple_length}
\title{Returns tuple length.}
\usage{
code_tuple_length(code, length = -1L)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
Number tuple length.
}
\description{
Returns the tuple length of the code. If the code is a mixed code it returns the longest tuple length.
}
\examples{
code_l <- code_tuple_length(c("ACG", "CAG"))
code_l <- code_tuple_length("ACGCAG", 3)
code_l <- code_tuple_length("ACG CAG")

}
