% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{codons_to_tessera}
\alias{codons_to_tessera}
\title{Tessera from codons}
\usage{
codons_to_tessera(code)
}
\arguments{
\item{code}{is either a string vector or a string. It can either be a code or a sequence.}
}
\value{
the argument code transfomed to a set of tesserae
}
\description{
This function uses a transformation to map all codons to a tessera. This transformation was published by Gonzalez, Giannerini and Rosa.
}
\examples{
tessera <- codons_to_tessera(c("ACG", "GAT"))
tessera <- codons_to_tessera("ACGGAT")
tessera <- codons_to_tessera("ACG GAT")

}
