% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_amd_analysis_code_in_sequence}
\alias{find_amd_analysis_code_in_sequence}
\title{Analysis a sequence based on a generic code.}
\usage{
find_amd_analysis_code_in_sequence(seq, code, tuple_length = -1L)
}
\arguments{
\item{code}{is either a string vector or a string. it can be a DNA or RNA sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
Returns a List with all analysing results. The list contains following:\cr
\emph{words} all found words of the code in the sequence in the correct order.\cr
\emph{idx_list} the first-letter index of all found words of the code in the sequence in the correct order.\cr
\emph{rest} all parts of the sequence which are not matching the code.\cr
\emph{parts} the sequence separated in matching and non matching parts. Odd indexes are matching, even indexes are not matching.\cr
\emph{longest_match} the longest connected matching sequence.\cr
\emph{total_match_in_percent} the percentage of the matching parts.
}
\description{
Finds all appearance of a code ion a sequence. Finds the longest connected matches of words of the code in the sequence.
The function also calculates the matching bases of the found words in the sequence.
}
\examples{
seq <- "ACGTCGCGACGTACGACGTCGTACTCGATGCAAGATC"
res <- find_amd_analysis_code_in_sequence(seq, c("ACG", "TCG"))
res <- find_amd_analysis_code_in_sequence(seq, "ACGCG", tuple_length=3)
res <- find_amd_analysis_code_in_sequence(seq, "ACG TCG")

}
