% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_code_by_min_value}
\alias{generate_code_by_min_value}
\title{This function generates a maximal circular code.}
\usage{
generate_code_by_min_value(alphabet, tuple_length)
}
\arguments{
\item{alphabet}{is a vector of letters.}

\item{tuple_length}{is the tuple length of the resulting code.}
}
\value{
A maximal circular code as String vector.
}
\description{
Based on a value function each word can be assigned to a value. The code is a set of the words with the
lowest value of each permutation class. This code has be proven to be circular and maximal. This can be done for any alphabet.
}
\examples{
code <- generate_code_by_min_value(c("A", "C", "G", "T"), 3)

}
