% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgorithm.R
\name{generate_weight_matrix}
\alias{generate_weight_matrix}
\title{Generate a default weight matrix list}
\usage{
generate_weight_matrix(ca_obj, tuple_length = 3, alphabet = NULL)
}
\description{
Generate a default weight matrix list
}
\examples{
# default tuple length is 3, default alphabet is RNA
w.m <- generate_weight_matrix()
# w.m is:
# [[1]]
#   T C A G
# T 1 1 1 1
# C 1 1 1 1
# A 1 1 1 1
# G 1 1 1 1
# 
# [[2]]
#   T C A G
# T 1 1 1 1
# C 1 1 1 1
# A 1 1 1 1
# G 1 1 1 1
# 
# [[3]]
#   T C A G
# T 1 1 1 1
# C 1 1 1 1
# A 1 1 1 1
# G 1 1 1 1

}
