% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genetic_codes_by_name}
\alias{genetic_codes_by_name}
\title{The Genetic Codes}
\usage{
genetic_codes_by_name(name)
}
\arguments{
\item{name}{the name of a Genetic Code as string. (check \link{print_all_translation_table})}
}
\value{
Returns a named List with all codons and the translated amino acids:\cr
}
\description{
\emph{genetic_codes_by_name} returns a list representing a Genetic Code. The list contains two arrays, one array of the codons and one
array of the translated amino acids in the same order.\cr
Source:\cr
\emph{Compiled by Andrzej (Anjay) Elzanowski and Jim Ostell at National Center for Biotechnology Information (NCBI), Bethesda, Maryland, U.S.A.}\cr
\emph{Last update of the Genetic Codes: Nov. 18, 2016}
\emph{\link{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi}}
}
\details{
\emph{codons} all codons as strings.\cr
\emph{amino_acids} the translated aminop acids in same order.\cr
}
\examples{
(code <- genetic_codes_by_name("The Yeast Mitochondrial Code"))

}
