% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_all_rna_tuples}
\alias{get_all_rna_tuples}
\title{Tuple Generator: get_all_rna_tuples}
\usage{
get_all_rna_tuples(tuple_length)
}
\arguments{
\item{tuple_length}{a number, the length of the block or tuple used}
}
\value{
A lexically sorted list of all tuples over the RNA Alphabet of length tuple_size\cr
}
\description{
Generats a lexically sorted list of all tuples over the RNA Alphabet of length tuple_size
}
\examples{
res <- get_all_rna_tuples(3)

}
