% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonTables.R
\name{get_dna_codon_list}
\alias{get_dna_codon_list}
\title{A List of all DNA codons}
\usage{
get_dna_codon_list()
}
\value{
A vector object with all DNA codons
}
\description{
get_dna_codon_list will generate the following list:\cr
TTT TCT TAT TGT TTC TCC TAC TGC TTA TCA TAA TGA TTG TCG TAG TGG CTT CCT CAT CGT CTC CCC CAC CGC CTA CCA CAA CGA CTG CCG CAG CGG
ATT ACT AAT AGT ATC ACC AAC AGC ATA ACA AAA AGA ATG ACG AAG AGG GTT GCT GAT GGT GTC GCC GAC GGC GTA GCA GAA GGA GTG GCG GAG GGG
}
\examples{
res <- get_dna_codon_list()

}
