% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonTables.R
\name{get_dna_codon_table}
\alias{get_dna_codon_table}
\title{A Matrix of all DNA codons}
\usage{
get_dna_codon_table()
}
\value{
A matrix object with all DNA codons
}
\description{
get_dna_codon_table will generate a 4*16 table based on the following list:\cr
 TTT TCT TAT TGT\cr
 TTC TCC TAC TGC\cr
 TTA TCA TAA TGA\cr
 TTG TCG TAG TGG\cr
 CTT CCT CAT CGT\cr
 CTC CCC CAC CGC\cr
 CTA CCA CAA CGA\cr
 CTG CCG CAG CGG\cr
 ATT ACT AAT AGT\cr
 ATC ACC AAC AGC\cr
 ATA ACA AAA AGA\cr
 ATG ACG AAG AGG\cr
 GTT GCT GAT GGT\cr
 GTC GCC GAC GGC\cr
 GTA GCA GAA GGA\cr
 GTG GCG GAG GGG
}
\examples{
res <- get_rna_codon_table()

}
