% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_max_conductance_of_codeidx}
\alias{get_max_conductance_of_codeidx}
\title{To calculate the max conductance of a translation table}
\usage{
get_max_conductance_of_codeidx(codeIdx, acid = "DNA")
}
\arguments{
\item{codeIdx}{a Number, the index of the genetic translation table. (see \link{print_all_translation_tables})}

\item{acid}{a String, is optional: DNA or RNA}
}
\value{
A Number, the max conductance of a genetic translation table
}
\description{
This function calculates the max conductance for a selected genetic translation table.
The conductance is a number which reflects the robustness of a genetic translation table.
The conductance the ration of the codons translating one amino acid which differ in only one position and the
codons translating the same amino acid but differ in more then one position.
}
\examples{
ac = get_average_conductance_of_code("The Standard Code")
ac = get_average_conductance_of_code("The Vertebrate Mitochondrial Code", "RNA")
}
