% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonTables.R
\name{get_rna_codon_list}
\alias{get_rna_codon_list}
\title{A List of all RNA codons}
\usage{
get_rna_codon_list()
}
\value{
A vector object with all RNA codons
}
\description{
get_rna_codon_list will generate the following list:\cr
UUU UCU UAU UGU UUC UCC UAC UGC UUA UCA UAA UGA UUG UCG UAG UGG CUU CCU CAU CGU CUC CCC CAC CGC CUA CCA CAA CGA CUG CCG CAG CGG
AUU ACU AAU AGU AUC ACC AAC AGC AUA ACA AAA AGA AUG ACG AAG AGG GUU GCU GAU GGU GUC GCC GAC GGC GUA GCA GAA GGA GUG GCG GAG GGG
}
\examples{
res <- get_rna_codon_list()

}
