% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonTables.R
\name{get_rna_codon_table}
\alias{get_rna_codon_table}
\title{A Matrix of all RNA codons}
\usage{
get_rna_codon_table()
}
\value{
A matrix object with all RNA codons
}
\description{
get_rna_codon_table will generate a 4*16 table based on the following list:\cr
UUU UCU UAU UGU\cr
AUU ACU AAU AGU\cr
}
\examples{
res <- get_rna_codon_table()

}
