% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_all_translation_table}
\alias{print_all_translation_table}
\title{The Genetic Codes}
\usage{
print_all_translation_table()
}
\description{
\emph{print_all_translation_table} Prints a list of the genetic codes containt by this project. The list includes names and indexes of the codes.
}
\details{
The Standard Code(transl_table=1)\cr
The Vertebrate Mitochondrial Code(transl_table=2)\cr
The Yeast Mitochondrial Code(transl_table=3)\cr
The Mold, Protozoan, and Coelenterate Mitochondrial Code and the Mycoplasma/Spiroplasma Code(transl_table=4)\cr
The Invertebrate Mitochondrial Code(transl_table=5)\cr
The Ciliate, Dasycladacean and Hexamita Nuclear Code(transl_table=6)\cr
The Echinoderm and Flatworm Mitochondrial Code(transl_table=9)\cr
The Euplotid Nuclear Code(transl_table=10)\cr
The Bacterial, Archaeal and Plant Plastid Code(transl_table=11)\cr
The Alternative Yeast Nuclear Code(transl_table=12)\cr
The Ascidian Mitochondrial Code(transl_table=13)\cr
The Alternative Flatworm Mitochondrial Code(transl_table=14)\cr
Ter Chlorophycean Mitochondrial Code(transl_table=16)\cr
Trematode Mitochondrial Code(transl_table=21)\cr
Scenedesmus obliquus Mitochondrial Code(transl_table=22)\cr
Thraustochytrium Mitochondrial Code(transl_table=23)\cr
Pterobranchia Mitochondrial Code(transl_table=24)\cr
Candidate Division SR1 and Gracilibacteria Code(transl_table=25)\cr
Pachysolen tannophilus Nuclear Code(transl_table=26)\cr
Karyorelict Nuclear(transl_table=27)\cr
Condylostoma Nuclear(transl_table=28)\cr
Mesodinium Nuclear(transl_table=29)\cr
Peritrich Nuclear(transl_table=30)\cr
Blastocrithidia Nuclear(transl_table=31)\cr
\cr Source:\cr
\emph{Compiled by Andrzej (Anjay) Elzanowski and Jim Ostell at National Center for Biotechnology Information (NCBI), Bethesda, Maryland, U.S.A.}\cr
\emph{Last update of the Genetic Codes: Nov. 18, 2016}
\emph{\link{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi}}
}
\seealso{
\link{genetic_codes_by_index}, \link{genetic_codes_by_name}
}
