% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_bda}
\alias{run_bda}
\title{Execute BDA procedure.}
\usage{
run_bda()
}
\value{
A StringVector. A binary representation of the Code.
}
\description{
Run the currently running BDA procedure. Therefore, you need to start a BDA procedure. 
You can call \code{\link{code_start_bda}} or \code{\link{seq_start_bda}} to start the procedure for a code. 
To add a rule simply Call \code{\link{add_bda_rule}}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
code_start_bda(c("ACG", "CAG"))
add_bda_rule(1, 2, "C", "G", "A", "T")
res.binary <- run_bda()

}
