% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdaTools.R
\name{run_bda_as_matrix}
\alias{run_bda_as_matrix}
\title{Execute BDA procedure (only works for all codons).}
\usage{
run_bda_as_matrix()
}
\value{
A matrix representation of the BDA procedure results.
}
\description{
Run the currently open BDA procedure. Therefore, you need to start a BDA procedure. For this method the procedure has to be started by the \code{\link{start_bda}} function. To start a procedure
for all codons call \code{\link{start_bda}}.
To add a BDA simply Call \code{\link{add_bda}}. To execute the BDAs call \code{\link{run_bda_as_matrix}} or \code{\link{run_bda}}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
start_bda()
add_bda(1, 2, "C", "G", "A", "U")
add_bda(1, 3, "A", "G", "C", "U")
res <- run_bda_as_matrix()

}
