% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_bda_for_all_rna_codons}
\alias{run_bda_for_all_rna_codons}
\title{Execute BDA procedure for all codons.}
\usage{
run_bda_for_all_rna_codons()
}
\value{
A List. A binary representation of the Code.
}
\description{
This function executes the currently open binary dichotomic algorithm (BDA) procedure. Therefore, you need to start a BDA procedure.
You can call \link{code_start_bda} to start a procedure
for all codons call \link{start_bda}.
To add a BDA simply Call \link{add_bda}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
code_start_bda(c("ACG", "CAG"))
add_bda(1, 2, "C", "G", "A", "T")
add_bda(1, 3, "A", "G", "C", "T")
res.binary <- run_bda_for_all_codons()

}
\seealso{
[start_bda()] for all codons, [add_bda] add a BDA,
          [code_start_bda()] for code
}
