% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPloter.R
\name{seq_factor_c3graph}
\alias{seq_factor_c3graph}
\title{Prepares a graph G(X) object with C3 extension.}
\usage{
seq_factor_c3graph(seq)
}
\arguments{
\item{seq}{A gene sequence as string.}

\item{show_circles}{A bool value. If true the all edges building a circle are colored red.}

\item{show_longest_path}{A bool value. If true the all edges part of the longest path are colored blue.}
}
\description{
Prepares a R igraph object. Extents the Graph G(X) from code_factor_graph by the so called C3 edges.
If the graph shows a cycle with alternating edges of the C3 edges and the usual edges follows that the
at least one circular permutation of the code is not a circular code. Only works for word length of 3.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
G <- seq_factor_c3graph("ACGCGA", )
plot(G)

}
\seealso{
[code_factor_c3graph(code)] for code
}
