% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPloter.R
\name{seq_factor_circles}
\alias{seq_factor_circles}
\title{Prepares a graph G(X) with only circle paths.}
\usage{
seq_factor_circles(seq, word_length = 3)
}
\arguments{
\item{code}{A DNA or RNA code as string vector represented by the graph.}
}
\description{
Prepares a R igraph object. Extracts all circle path of the Graph G(X).
If the graph shows no cycle the graph will be empty. Otherwise it returns a igraph object showing all
circle paths.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
G <- seq_factor_circles(c("ACG", "CGA"))
plot(G)

}
\seealso{
[code_factor_circles(code)] for code, [code_get_all_circle_as_vector()] for the results as vector
}
