% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPloter.R
\name{seq_factor_graph}
\alias{seq_factor_graph}
\title{Prepares a R igraph object.}
\usage{
seq_factor_graph(seq, word_length = 3, show_circles = FALSE,
  show_longest_path = FALSE)
}
\arguments{
\item{seq}{A gene sequence as string.}

\item{word_length}{A number, the length of the codons in the seq.}

\item{show_circles}{A bool value. If true the all edges building a circle are colored red.}

\item{show_longest_path}{A bool value. If true the all edges part of the longest path are colored blue.}
}
\description{
The funciton prepares graph G(X) object for a sequence. The following definition relates a directed graph to
any n-nucleotide code. Recall from graph theory (Clark and Holton, 1991) that a graph G consists of
a finite set of vertices (nodes) V and a finite set of edges E. Here, an edge is a set \{v,w\} of vertices
from V . The graph is called oriented if the edges have an orientation, i.e. edges are considered to be
ordered pairs [v,w] in this case.\cr
Definition 2.1. Let X Bn be an n-nucleotide code (n 2 N). We define a directed graph G(X) =
(V (X),E(X)) with set of vertices V (X) and set of edges E(X) as follows:
N-NUCLEOTIDE CIRCULAR CODES IN GRAPH THEORY 5\cr
V (X) = \{N1...Ni,Ni+1...Nn : N1N2N3...Nn in X, 0 < i < n\}\cr
E(X) = \{[N1...Ni,Ni+1...Nn] : N1N2N3...Nn in X, 0 < i < n\}\cr
The graph G(X) is called the representing graph of X or the graph associated to X.\cr
Basically, the graph G(X) associated to a code X interprets n-nucleotide words from X in (n−1) ways
by pairs of i-nucleotides and (n-i)-nucleotides for 0 < i < n.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
G <- seq_factor_graph("ACGCGA", 3)
plot(G)

}
\seealso{
[code_factor_graph(code)] for code
}
