% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_get_info}
\alias{seq_get_info}
\title{Sequence, Code analyzer}
\usage{
seq_get_info(seq, tuple_length = 3L)
}
\arguments{
\item{tuple_length}{a number, the length of the block or tuple used}

\item{a}{character string a sequence of letters and/or numbers}
}
\value{
A named list with the listed values:\cr
\emph{alphabet} (String) All letters and symbols used.\cr
\emph{number_of_tuple} (Number) The number of all words/tuples used.\cr
\emph{tuple_count} (Key value map) Key value map of the words in the sequence.\cr
}
\description{
Returns a table with all words in a sequence. For each word, the table shows how often the word appears in the sequence.
}
\examples{
res <- seq_get_info("ACGCGAACG", 3)

}
