% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_prepare_plot_longest_path}
\alias{seq_prepare_plot_longest_path}
\title{Returns a list of circles as vectors.}
\usage{
seq_prepare_plot_longest_path(seq, word_length)
}
\arguments{
\item{seq}{a nucleotide sequence.}

\item{word_length}{the length of the word.}
}
\description{
The function checks if the code is circular. If the code is not circular the functions returns all circles. 
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
seq_prepare_plot_all_circles("ACGCAG", 3)

}
