% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_start_bda}
\alias{seq_start_bda}
\title{Start a BDA procedure for a sequence.}
\usage{
seq_start_bda(seq, word_length)
}
\arguments{
\item{seq}{A gene sequence as string.}

\item{word_length}{A number, the length of the codons in the seq.}
}
\value{
Boolean value. True if the procedure has started
}
\description{
To generate a BDA based binary representation of a code or a list of codons, you need to start a BDA procedure. 
To do so you can eihter call \code{\link{code_start_bda}} or \code{\link{seq_start_bda}}. To start a procedure
for all codons call \code{\link{start_bda}}.
To add a BDA simply Call \code{\link{add_bda}}. To execute the BDAs call \code{\link{run_bda}}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
seq_start_bda("ACGCAG", 3)
add_bda(1, 2, "C", "G", "A", "T")
add_bda(1, 3, "A", "G", "C", "T")
res <- run_bda()

}
\seealso{
[code_start_bda()] for code, [start_bda()] for all codons, [add_bda] add a BDA,
          [run_bda] for executing and run the BDAs
}
