% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shift_tuples}
\alias{shift_tuples}
\title{Circular shift of all tuples}
\usage{
shift_tuples(shifts, code, tuple_length = -55555L)
}
\arguments{
\item{shifts}{number of shifts}

\item{code}{is either a string vector or a string. It can either be a code or a sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
shifted code as String vector
}
\description{
This functions shifts all tuples in code anticlockwise. In other words, the first character of each tuples gets removed and
added to the end of the same tuple. Depending on the parameter \emph{shift} this process is repeated multiple times.
}
\examples{
shifted_code <- shift_tuples(2, c("ACG", "GAT"))
shifted_code <- shift_tuples(2, "ACGGAT", tuple_length=3)
shifted_code <- shift_tuples(2, "ACG GAT")

}
