% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BdaTools.R
\name{start_bda}
\alias{start_bda}
\title{Start a BDA procedure for all codons}
\usage{
start_bda()
}
\value{
Boolean value. True if the procedure has started
}
\description{
To generate a BDA based binary representation of a code or a list of codons, you need to start a BDA procedure. 
To do so you can eihter call \code{\link{code_start_bda}} or \code{\link{seq_start_bda}}. To start a procedure
for all RNA codons call \code{\link{start_bda}}.
To add a BDA simply Call \code{\link{add_bda}}. To execute the BDAs call \code{\link{run_bda_as_matrix}} or \code{\link{run_bda}}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
start_bda()
add_bda(1, 2, "C", "G", "A", "U")
add_bda(1, 3, "A", "G", "C", "U")
res <- run_bda_as_matrix()

}
