% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgorithm.R
\name{start_cluster_algorithm}
\alias{start_cluster_algorithm}
\title{Start a Cluster Algorithm object}
\usage{
start_cluster_algorithm(code)
}
\arguments{
\item{code}{is a string vector. It is an alternating list of code words and their translations.}
}
\description{
Use this object to run more advanced methods.
}
\examples{
# AC -> 01 & AA -> 01 &  CA -> 11 & CC -> 10
ca_obj <-  start_cluster_algorithm(c("AC", "01", "AA", "01", "CA", "11", "CC", "10"))

}
