% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgorithm.R
\name{start_cluster_algorithm_of_codeidx}
\alias{start_cluster_algorithm_of_codeidx}
\title{Start a Cluster Algorithm object}
\usage{
start_cluster_algorithm_of_codeidx(codeIdx = 1)
}
\arguments{
\item{codeIdx}{a Number, the index of the genetic translation table. (see \link{print_all_translation_tables})}
}
\description{
This function starts a cluster algorithm for a selected genetic translation table.
The conductance is a number which reflects the robustness of a genetic translation table.
The conductance the ration of the codons translating one amino acid which differ in only one position and the
codons translating the same amino acid but differ in more then one position.
}
\examples{
# For the SGC
ca_obj <-  start_cluster_algorithm_of_codeidx(1)

}
