% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAlgorithm.R
\name{start_cluster_algorithm_on_bda}
\alias{start_cluster_algorithm_on_bda}
\title{Start a Cluster Algorithm object}
\usage{
start_cluster_algorithm_on_bda(bda_obj)
}
\arguments{
\item{bda_obj}{BDA object (see \link{BDATools})}
}
\description{
This function starts a cluster algorithm for a BDA.
The conductance is a number which reflects the robustness of a genetic translation table.
The conductance the ration of the codons translating one amino acid which differ in only one position and the
codons translating the same amino acid but differ in more then one position.
}
\examples{
# For a simple BDA
bda_obj<-start_bda()
add_rule(bda_obj, 1, 2, "C", "G", "A", "U")
ca_obj <-  start_cluster_algorithm_on_bda(bda_obj)

}
