% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tessera_to_codons}
\alias{tessera_to_codons}
\title{Codons from Tessera}
\usage{
tessera_to_codons(code)
}
\arguments{
\item{code}{is either a  tessera string vector or a string. It can either be a code or a sequence.}
}
\value{
the argument code transfomed to a set of codons
}
\description{
This function uses a transformation to map all tessera to a codon. This transformation was published by Gonzalez, Giannerini and Rosa.
}
\examples{
codons <- codons_to_tessera(c("ACGT", "GATC"))
codons <- codons_to_tessera("ACGTGATC")
codons <- codons_to_tessera("ACGT GATC")

}
