% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tessera_to_table_str}
\alias{tessera_to_table_str}
\title{Tables associated to Tessera as String}
\usage{
tessera_to_table_str(code)
}
\arguments{
\item{code}{is either a  tessera string vector or a string. It can either be a code or a sequence.}
}
\value{
A 4*4 tabele associated to the Tessera code as String
}
\description{
This function generates a table that represents a Tessera code. The exact definition can be found in the article: "Circular Tessera Codes in the Evolution of the Genetic Code".
}
\examples{
table <- tessera_to_table_str(c("ACGT", "GATC"))
table <- tessera_to_table_str("ACGT GATC")

}
