% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_block_length}
\alias{code_block_length}
\title{Block length}
\usage{
code_block_length(code)
}
\arguments{
\item{code}{is either a string vector or a string. It has to be a set of words/blocks}
}
\value{
block length
}
\description{
This function gets the block length of a code.
}
\examples{
block.length <- code_block_length(c("ACG", "GAT"))
block.length <- code_block_length("ACG GAT")

}
