% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_check_if_cn_circular}
\alias{code_check_if_cn_circular}
\title{Check if a code is C_{n} circular.}
\usage{
code_check_if_cn_circular(code)
}
\arguments{
\item{code}{A vector with codons.}
}
\value{
Boolean vector. True if the code C_{n} circular.
}
\description{
This function checks the same property as \code{\link{check_if_circular}}. 
Yet,  this function checks the property for for all circular permutations of the codons in the code.
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_check_if_cn_circular(c("ACG", "CAG"))

}
