% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_check_if_self_complementary}
\alias{code_check_if_self_complementary}
\title{Check if a code is self complementary.}
\usage{
code_check_if_self_complementary(code, length = -1L, mute = TRUE)
}
\arguments{
\item{code}{is either a string vector or a string. It has to be a RNA/DNA - code or a sequence.}

\item{length}{if code is a sequence, length is the tuple length of the code.}

\item{mute}{set false to get console output information about not self-complementary tuples.}
}
\value{
Boolean value. True if the code is self-complementary.
}
\description{
This function checks if a code is self complementary.
The code can either be a vector of DNA/RNA tuples or a sequence. If the code
is a sequence an additional word length parameter is needed.\cr
A code is self complementary if and only if for all tuples in the code the anti-tuple is also in the code.
An anti-tuple is te reversed tuple of complementary bases.\cr A <-> T (U) and C <-> G. The anti-tuple of ACG is CGT
\link{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5492142/},\cr
\link{http://dpt-info.u-strasbg.fr/~c.michel/Circular_Codes.pdf},\cr
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_check_if_self_complementary(c("ACG", "CAG"))
code_check_if_self_complementary("ACGCAG", 3)
code_check_if_self_complementary("ACG CAG")

}
