% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPloter.R
\name{code_factor_c3graph}
\alias{code_factor_c3graph}
\title{Prepares a graph G(X) object with C3 extension.}
\usage{
code_factor_c3graph(code)
}
\arguments{
\item{code}{A DNA or RNA code as string vector represented by the graph.}
}
\description{
Prepares a R igraph object. Extents the Graph G(X) from code_factor_graph by the so called C3 edges.
If the graph shows a cycle with alternating edges of the C3 edges and the usual edges follows that the
at least one circular permutation of the code is not a circular code. Only works for word length of 3.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
G <- code_prepare_factor_gen_c3graph("ACGCGA", )
plot(G)

}
\seealso{
[seq_factor_c3graph(seq)] for sequences
}
