% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPloter.R
\name{code_factor_longest_path}
\alias{code_factor_longest_path}
\title{Prepares a graph G(X) with only longest paths.}
\usage{
code_factor_longest_path(code, tuple_length = -1)
}
\arguments{
\item{code}{is either a string vector or a string. it can be a DNA or RNA sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\description{
Prepares a R igraph object. Extracts all longest path of the Graph G(X).
If the graph shows a cycle the graph will be empty. Otherwise it returns a igraph object showing all
longest paths.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
G <- code_factor_longest_path(c("ACG", "CAG"))
G <- code_factor_longest_path("ACGCAG", tuple_length=3)
G <- code_factor_longest_path("ACG CAG")
plot(G)

}
\seealso{
[code_factor_circles(code)] for code, [code_get_all_longest_path_as_vector()] for the results as vector
}
