% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_get_all_circle_as_vector}
\alias{code_get_all_circle_as_vector}
\title{Returns a list of circles as vectors.}
\usage{
code_get_all_circle_as_vector(code)
}
\arguments{
\item{code}{A vertor with codons.}
}
\value{
List of string vectors.
}
\description{
The function checks if the code is circular. If the code is not circular the functions returns all circles. 
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_get_all_circle_as_vector(c("ACG", "CAG"))

}
