% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_get_one_circle_as_vector}
\alias{code_get_one_circle_as_vector}
\title{Finds one circular sequences constructable in a code.}
\usage{
code_get_one_circle_as_vector(code, tuple_length = -1L)
}
\arguments{
\item{code}{is either a string vector or a string. it can be a DNA or RNA sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
List of on circular sequences constructable with a given code.
}
\description{
Prepares a R path string vector. Extracts all circular paths of the Graph G(X).
If the graph shows no cycles the vector will be empty. Otherwise it returns a vector with one
circular path.\cr
\emph{2007 E. FIMMEL, C. J. MICHEL, AND L. STRÜNGMANN. N-nucleotide circular codes in graph theory}
}
\examples{
l_graph <- code_get_one_circle_as_vector(c("ACG", "CGA"))
l_graph <- code_get_one_circle_as_vector("ACGCGA", 3)
l_graph <- code_get_one_circle_as_vector("ACG CGA")

}
\seealso{
[code_factor_circles(code)] for code, [code_get_all_circle_as_vector()] for the results as list
}
