% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_prepare_factor_all_cycle}
\alias{code_prepare_factor_all_cycle}
\title{Returns a list, with only the cycles in the graph of a circular code.}
\usage{
code_prepare_factor_all_cycle(code, tuple_length = -55555L)
}
\arguments{
\item{code}{is either a string vector or a string. It can be a DNA or RNA sequence.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
List: Edges and vertices of only the cycles of an graph representing a circular code.
}
\description{
This function is based on the represent graph of a circular code (see \link{code_prepare_factor_graph}).
The function only returns a list of edges and vertices. To get a graph object use \link{code_factor_cycle}.
The function checks if the code is circular. If the code is not circular the functions returns all cycles in the representing graph.
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_prepare_factor_all_cycle(c("ACG", "CAG"))
code_prepare_factor_all_cycle("ACGCAG", tuple_length=3)
code_prepare_factor_all_cycle("ACG CAG")

}
