% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_prepare_factor_longest_path}
\alias{code_prepare_factor_longest_path}
\title{Returns a list of circles as vectors.}
\usage{
code_prepare_factor_longest_path(code)
}
\arguments{
\item{code}{A vertor with codons.}
}
\description{
The function checks if the code is circular. If the code is not circular the functions returns all circles. 
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_prepare_factor_all_circles(c("ACG", "CAG"))

}
