% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_prepare_factor_longest_path}
\alias{code_prepare_factor_longest_path}
\title{Returns a list, with only the longest path(s) in the graph of a circular code.}
\usage{
code_prepare_factor_longest_path(code, tuple_length = -1L)
}
\arguments{
\item{code}{A vertor with codons.' @param length if code is a sequence, length is the tuple length of the code.}

\item{tuple_length}{if code is a sequence, length is the tuple length of the code.}
}
\value{
List: Edges and vertices of only the longest path of an graph representing a circular code.
}
\description{
This function is based on the represent graph of a circular code (see \link{code_prepare_factor_graph}).
The function only returns a list of edges and vertices. To get a graph object use \link{code_factor_longest_path}.
The function checks if the code is circular. If the code is not circular the functions returns the logest paths.
\emph{2007 Christian MICHEL. CIRCULAR CODES IN GENES}
}
\examples{
code_prepare_factor_longest_path(c("ACG", "CAG"))
code_prepare_factor_longest_path("ACGCAG", tuple_length=3)
code_prepare_factor_longest_path("ACG CAG")

}
