% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_start_bda}
\alias{code_start_bda}
\title{Start a BDA procedure for a code.}
\usage{
code_start_bda(code)
}
\arguments{
\item{code}{A vertor with codons.}
}
\value{
Boolean value. True if the procedure has started
}
\description{
To generate a BDA based binary representation of a code or a list of codons, you need to start a BDA procedure. 
To do so you can eihter call \code{\link{code_start_bda}} or \code{\link{seq_start_bda}}. To start a procedure
for all codons call \code{\link{start_bda}}.
To add a BDA simply Call \code{\link{add_bda}}. To execute the BDAs call \code{\link{run_bda}}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
code_start_bda(c("ACG", "CAG"))
add_bda(1, 2, "C", "G", "A", "T")
res.binary <- run_bda()

}
