% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{code_start_bda}
\alias{code_start_bda}
\title{Start a BDA procedure.}
\usage{
code_start_bda(code)
}
\arguments{
\item{code}{A vertor with codons.}
}
\value{
Boolean value. True if the procedure has started
}
\description{
A binary dichotomic algorithm generates a binary representation of a generic code. To add a rule simply Call \code{\link{add_bda_rule}}.
For more info on BDAs read:
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic Code generated by Binary Dichotomic Algorithms}
}
\examples{
code_start_bda(c("ACG", "CAG"))
add_bda_rule(1, 2, "C", "G", "A", "T")
res.binary <- run_bda()

}
