% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_bda}
\alias{run_bda}
\title{Execute BDA procedure.}
\usage{
run_bda()
}
\value{
A List. A binary representation of the Code.
}
\description{
Run the currently open BDA procedure. Therefore, you need to start a BDA procedure. 
You can call \code{\link{code_start_bda}} or \code{\link{seq_start_bda}} for a code. To start a procedure
for all codons call \code{\link{start_bda}}.
To add a BDA simply Call \code{\link{add_bda}}. To execute the BDAs call \code{\link{run_bda}}.
A binary dichotomic algorithm generates a binary representation of a generic code. For more info on BDAs read:\cr
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic  Code generated by Binary Dichotomic Algorithms}
}
\examples{
code_start_bda(c("ACG", "CAG"))
add_bda(1, 2, "C", "G", "A", "T")
add_bda(1, 3, "A", "G", "C", "T")
res.binary <- run_bda()

}
\seealso{
[seq_start_bda()] for sequence, [start_bda()] for all codons, [add_bda] add a BDA,
          [code_start_bda()] for code
}
