% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seq_start_bda}
\alias{seq_start_bda}
\title{Start a BDA procedure.}
\usage{
seq_start_bda(seq, word_lenth)
}
\arguments{
\item{seq}{A gene sequence.}

\item{word_lenth}{length of the codons.}
}
\value{
Boolean value. True if the procedure has started
}
\description{
A binary dichotomic algorithm generates a binary representation of a generic code. To add a rule simply Call \code{\link{add_bda_rule}}.
For more info on BDAs read:
\emph{M.Gumbel, E.Fimmel, A.Danielli, L.Struengmann. On Models of the Genetic Code generated by Binary Dichotomic Algorithms}
}
\examples{
seq_start_bda("ACGCAG", 3)
add_bda_rule(1, 2, "C", "G", "A", "T")
res.binary <- run_bda()

}
